+++ README.md
@@ -1,1 +1,16 @@
-# sprint_5
+# Тесты на сервис "Доска"
+
+## Структура тестов
+
+- **`test_registration_page`** — тесты на форму регистрации
+- **`test_make_announcement`** — тесты на создание объявлений
+- **`test_login_page`** — тест на авторизацию пользователя
+- **`test_logout_page`** — тест на выход из личного кабинета
+
+## Дополнительные модули
+
+- **`ui_helpers`** — вспомогательные методы для работы с UI
+- **`locators`** — используемые локаторы, разбиты по классам для удобства чтения
+- **`data`** — вспомогательные данные (например, тестовые пользователи, тексты, конфиги)
+- **`conftest`** — фикстуры для тестов
+
